separate Notification from NavBar
